+++ VegaEngine2/VegaEngine2/src/Utility/UI/VegaUI.h
@@ -10,7 +10,7 @@ namespace fz {
 		static bool DrawControl1(const std::string& label, const std::string& buttonLabel,
 								 int& v,
 								 int speed_x = 1,
-								 int x_min = 0, float x_max = 0,
+								 int x_min = 0, int x_max = 0,
 								 int resetValue = 0, float columnWidth = 100.f);
 		static bool DrawControl1(const std::string& label, const std::string& buttonLabel,
 								 float& v,
